requirements.txt
fastapi
uvicorn
jinja2
python-multipart
nltk
requests
google-generativeai
python-dotenv
huggingface_hub
Pillow